\name{Banknotes}
\alias{Banknotes}
\title{Swiss banknotes data set}
\description{
The swiss banknotes data set contains 7 variables and 200 observations.
The first 6 variables are certain measures of explanatory variables of swiss banknotes, 
the last variable coded as 0/1 is the response variable. The last variable describes whether the banknote is genuine or not.
The data set is given in Riedwyl, H. (1997). Lineare Regression und Verwandtes. Birkhaeuser, Basel.
}
\usage{data(Banknotes)}
\format{
The swiss banknotes data set contains 7 variables and 200 observations.}
\keyword{datasets}

