\name{Z1}
\alias{Z1}
\title{Data set to illustrate the use of ncomplete.for}
\description{Data set to illustrate the use of ncomplete.for.}
\usage{data(Z1)}
\format{
The data set contains 2 columns and 10 rows. The first
column denotes an explanatory variable. The last column 
denotes the response variable coded as 0/1.}
\keyword{datasets}

