\name{Z2}
\alias{Z2}
\title{Data set to illustrate the use of ncomplete.for}
\description{
Data set to illustrate the use of ncomplete.for.
}
\usage{data(Z2)}
\format{
The data set contains 3 columns and 10 rows. The first
two columns denote two explanatory variables. The last column 
denotes the response variable coded as 0/1.}
\keyword{datasets}

